+++ src/lib/option-definitions.js
@@ -67,7 +67,7 @@ const optionDefinitions = {
     type: 'regexpArray'
   },
   customEventAttributes: {
-    description: 'Array of regexes that allow to support custom event attributes for minifyJS (e.g., `ng-click`)',
+    description: 'Array of regexes that allow to support custom event attributes (e.g., `ng-click`)—use with `--minify-js`',
     type: 'regexpArray'
   },
   decodeEntities: {
@@ -167,7 +167,7 @@ const optionDefinitions = {
     type: 'boolean'
   },
   removeEmptyElementsExcept: {
-    description: 'Array of elements to preserve when `--remove-empty-elements` is enabled (e.g., `td`, `<span aria-hidden="true">`)',
+    description: 'Array of elements to preserve (e.g., `td`, `<span aria-hidden="true">`)—use with `--remove-empty-elements`',
     type: 'jsonArray'
   },
   removeOptionalTags: {